requirements.txt
streamlit
sentence-transformers
pdfplumber
tiktoken
chromadb
python-dotenv
groq